fix: not install capsule error
